--- lib/INTERNAL.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = INTERNAL;
-
-function INTERNAL() {}
