+++ test/cover.js
@@ -2,7 +2,6 @@
 var aplus = require('promises-aplus-tests');
 var Promise = require('../lib');
 var adapter = {};
-var INTERNAL = require('../lib/INTERNAL');
 var assert = require('assert');
 
 adapter.deferred = function () {
@@ -13,10 +12,12 @@ adapter.deferred = function () {
   });
   return pending;
 };
-adapter.resolved = Promise.resolve;
-adapter.rejected = Promise.reject;
-//noop, just for coverage
-INTERNAL();
+adapter.resolved = function (value) {
+  return Promise.resolve(value);
+}
+adapter.rejected = function (reason) {
+  return Promise.reject(reason);
+}
 
 describe('Lie', function () {
   it('should work without new', function (done) {
